Add newsletter registration and Comments to Events
Incorporating what we learned in the last section about API routes, we added a POST request to a registration resource, which simply figures out the email the user typed in when registering and responds back with it. For comments, we simply hard-code a few comments and fetch them for the individual Event pages. We also allow the user to fill out a form and POST his own comment, though we don't really do anything with it.